添加了剩余逻辑（还有问题）

+++ ExamProject/UI/Controllers/EXExamineViewController.h
@@ -23,8 +23,8 @@
     NSTimer                     *_examTimer;        //倒计时用:除非推出或者提交，否则不可停止该倒计时器
     NSInteger                   _currentExamTime;
     BOOL                        _isExamSubmitted;
-    NSInteger                   _beginExamTime;
-    NSInteger                   _submitExamTime;
+    double                   _beginExamTime;
+    double                   _submitExamTime;
 }
 
 @property (nonatomic,retain)PaperData           *paperData;
